[refactor/#19] - instance와 interceptor 분리
